fix Int Node output type

+++ Assets/ProcGen/Runtime/Nodes/Constants/IntNode.cs
@@ -13,7 +13,7 @@ namespace ProcGen.Nodes.Constants
         }
         public override void Evaluate()
         {
-            Outputs[0].Value.Float = Value;
+            Outputs[0].Value.Int = Value;
         }
     }
 }
